Minor suggestions for CIGAR errors & pos parsing. (#43)
* Minor suggestions for CIGAR errors & pos parsing.

* Removed typo.

* TODO: Deprecate Cigar op Back in cigar.end_pos()?

* this implies changing the function's interface

* Adapt wording of include_del() to implementation.

The new wording reflects what testing of the code gave as the current
behaviour. @johanneskoester, is this the intended behaviour?

* Correct Cigar parsing for Pad and HardClip ops.

Ensure that HardClips are only allowed at read ends and Pads do not
consume any sequence (neither read nor reference). For the latter,
see this incoming clarification of the SAM specs:
https://github.com/jkbonfield/hts-specs/commit/0809a3074526f1f3c1f5e1f32a0a1314cb079ebb

* Tests for CigarStringView::read_pos().

* Deprecate Back ('B') in CigarStringView::end_pos()